+++ truffle/CHANGELOG.md
@@ -6,6 +6,7 @@ This changelog summarizes major changes between Truffle versions relevant to lan
 
 * Added `SuspendedEvent.setReturnValue` to change the return value of the currently executed source location.
 * Deprecated `FrameSlot#getIndex` without replacement.
+* Added `TruffleInstrument.Env.startServer()` to get a virtual message-based server provided via `MessageTransport` service.
 
 ## Version 1.0.0 RC7
 * Made all Truffle DSL annotations retention policy CLASS instead of RUNTIME. Reflecting DSL annotations at runtime is no longer possible. It is recommended to use `@Introspectable` instead.